Update rule imports
This commit updates how convolve2d is being imported inside the rules
module. In this case, we are importing scipy.signal rather than
importing the whole convolve2d itself.

The reason for this is that during tests, we just want to import all
exposed functions and run similar tests on them. When we do the latter
version we always get the convolve2d, and this might cause some
unexpected errors.

Signed-off-by: Lester James V. Miranda <ljvmiranda@gmail.com>